requirements_updated.txt
# خدمة الإدارة - مشروع نائبك
# متطلبات Python للإدارة والأمان والأداء

# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# File Handling & Images
Pillow==10.1.0
python-magic==0.4.27
django-storages==1.14.2
boto3==1.29.7

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-health-check==3.17.0

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.17.2
factory-boy==3.3.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
coverage==7.3.2

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
pre-commit==3.5.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Validation
marshmallow==3.20.1
jsonschema==4.20.0

# Performance
gunicorn==21.2.0
whitenoise==6.6.0
django-compressor==4.4

# Two-Factor Authentication
qrcode==7.4.2
pyotp==2.9.0

# Export & Reports
weasyprint==60.2
matplotlib==3.8.2
seaborn==0.13.0

# Environment
environs==10.0.0
